[LCR-352] refactor: conversion & questionnaire 폴더 분리
